Don't attempt to cancel tools without a canvas
This can happen when a file failed to open or a session didn't connect.
The tools don't expect to be asked to cancel when there's not even a
model for them to operate on, so that causes a crash. Now they're not
asked to cancel anything anymore in that case.