Use protocol version of correct hecking session
It just grabbed whatever protocol version the last session in the server
listing had. Which I guess in my cases always randomly happened to be a
beta session, because those are hosted later and so appear at the end of
the list. Now it actually uses the protocl version of the session you're
joining, which makes a whole lot more sense. Really stupid bug.

Fixes #1093.