Resolve ambiguous shortcuts with main window menu
When you assign (or a translation assigns) an Alt+Whatever shortcut to a
main window menu bar item that conflicts with a custom shortcut, it will
now resolve in favor of the custom shortcut instead of giving a bogus
message about an ambiguity you can't fix. It's still kinda random which
action you actually get, but at least it's clear what the problem is
that way.