Integrate Drawdance internally
Removing the subproject and the standalone stuff that's not currently
working, instead integrating Drawdance into Drawpile's sources directly.
It's no longer a separate project anyway, so we can get rid of the
complexity and IDE wonk that the subproject causes.

This also removes tests that are busted because of the protocol changing
as well as the engine benchmark code that's no longer in use.