Implement a start dialog
Instead of dumping the user into a blank document and expecting them to
fiddle with the window menu to get where they want. The dialog replaces
the join, browse, host and new dialogs, combining them into one unit. It
is also set to replace the new version check dialog, but that's not
implemented yet.