Remember last start dialog page for three days
Bringing you back to the previous page you were looking at, unless it's
been a while since you used Drawpile, in which case you're back on the
front page. The intent is that news and updates will also bring you back
to the front page, so doing it regularly should make that less jarring.