Clear local fork save points as they catch up
It pointlessly uses up memory to keep local fork savepoints in the
middle of strokes around, they're only useful if they're on or right
next to an undo point that may get rolled back to. So we now only keep
those useful states around and purge the ones that are in the middle of
nowhere and are very unlikely to serve as replay start points.