Make Fusion checkbox outlines more visible
When using a dark palette. The Fusion style has a really stupid value
hard-coded that makes this outline just meld entirely into most of the
background colors used in dark themes, so now we lighten that up harder
on dark themes.

This requires some really stupid hackery because the Fusion style is so
non-extensible, but it's necessary to make checkboxes in menus and tree
views visible at all.