Allow adding avatars in the login dialog
It happens frequently that new users join without knowing that they
could have given themselves an avatar. They ask how to do it in chat,
dig through the preferences to add it and then have to reconnect to make
it show up, which is just an annoying experience.

Now the login dialog gives an "add avatar" option, allowing users to
pick one right there. The dropdown to pick an avatar is of course now
also shown when the user doesn't have any avatars yet, rather than
hiding it in that case.

This also fixes up some other minor things about the avatar list model,
like making saving a bit more resilient and denoting which entries are
"real" and which ones are just the "no avatar" and "add avatar" options.