Store alpha preserve of a layer in ORA files
For compatibility with Krita and maybe other applications. In Drawpile,
alpha preserve is implicit depending on the blend mode, but e.g. in
Krita, it has to be enabled separately. This adds an
`alpha-preserve="true"` attribute to layers with affected blend mdoes so
that Krita can restore them properly, once the attribute is implemented
on their side.

Drawpile doesn't read the attribute, since it can't do anything with the
information, except for the special case of the Normal blend mode, which
becomes Recolor when alpha preserve is enabled. Normally this is saved
as a separate blend mode, but since it's a valid way to represent it, we
handle it too.