Bring back dprectool, dragging Rust along with it
It should work mostly the same, the command line flags are the same and
the behavior should also match up for the most part.

This reintroduces Rust into the project, only at the fringes for now, so
it can yet be avoided with -DTOOLS=OFF. But it will probably be
integrated into Drawdance to deal with stuff like file handling and
such, where using C is just terrible damage. The integration is less of
a back-and-forth mess than it was with Rustpile, since we no longer do
any linking in Rust space, we only produce static libraries in it and
let cmake deal with actually putting the executables together.