Switch to ahead-of-time bindgen generation
Because generating it during the build process doesn't work when
building Android and it's generally less annoying that way, since the
file is just out in the open instead of manifesting in some build
directory.

This excludes some system-specific stuff from the generated bindings,
such as endianness and the current system, since those are
host-specific.