Put brush preset database access behind a mutex
So that it doesn't get accessed from two threads at once and causes a
crash. This can happen on Android in some configurations. Also switches
addBindValue to bindValue with a position argument. Not sure if that is
related, but it works this way and we have all the indexes available
anyway, so might as well use them.

Relates to #1108 and #1140.