Let operators create layers beyond 256 of them
They were allowed to do this all along, the client just didn't attempt
to to so. It will now start creating layers owned by user 0 (the server)
and then create layers for user 255 and descending from there. This
doesn't remove the layer limit, but it allows for a workaround when e.g.
doing animations, you just need op permissions for it.