Unify leave and terminate on builtin server
Hitting Session > Leave will now ask if you want to leave and terminate
the session if you're the host on the builtin server, rather than just
asking if you want to leave and not mentioning that the session will be
obliterated. Session > Moderation > Terminate is now also available on
the builtin server, it just acts as if you clicked on Leave, since
shutting down the server is pretty terminal.