Replace reset notice dialog with on-canvas notices
The functionality is still the same, but the user interface is now less
disruptive. The bar showing the catchup progress is now shown on the
bottom-right corner of the canvas as a HUD element. The option to save
the canvas as it was before the reset is now on top of the canvas, where
it will automatically dismiss itself after 10 seconds if the user
continues to draw.