Make server messages translatable
They are sent to the client with a translation key now. If the client
knows the message, it will translate it, if not, it will continue to
show the literal message sent by the server. Old clients that don't have
the translation functionality also just show the message and ignore the
translation information.

Not all server messages are translatable yet, but the most important
ones should be.