Don't select layer when toggling its visibility
Mouse movements with the button held down cause the tree entry to be
selected, which is really annoying when you just want to hide a layer
and then it gets selected instead, making you have to go back and pick
the layer you were actually on. Now it makes sure not to let those mouse
move events after a visibility toggle get through to the layer tree.