Make catchup completion explicit
The server now sends a new caughtup message to clients when they're
actually caught up during a session reset or when joining a session. On
one hand, this fixes clients being allowed to draw too early, since
their catchup counter can be off due to messages received that aren't
from the session history. On the other, it resolves the issue where they
get stuck at 99% waiting for messages that never come and then it needs
a chat message or something to unstuck them. This can happen due to
invalid messages in the stream, which won't increase the counter.

Catchup messages now have a key as a correlator. Clients will only
consider a catchup done when they receive a caughtup message with a
matching key. The thin server sticks those caughtup messages into the
history, using 0 for session resets and rotating numbers from 1 to
999999 above that for clients joining. The builtin server just uses 0
for resets and 1 for clients catching up, since it doesn't need to store
those in the session history and so doesn't send caughtup messages to
clients that don't need them.

If the server doesn't send a correlation key along with their catchup
message, the client uses the old behavior of considering a catchup done
when it received enough messages.