Correct rounding error when converting 15 to 8 bit
It floored the values when using the AVX2 and SSE4.2 variants of those
functions, which caused some displayed colors to be off by a single bit
per channel. While that's not really perceptible, the color picker picks
those colors, so it would often drift by a tiny amount.

Now those functions round properly by adding a fudge factor to the
channel values. The "dumb" variants of those functions now also use the
same algorithm, since we might as well stay in integer space instead of
converting to floats and back.

There's a test added to verify that the values are the same as if using
the floating point route.