Fix crash on soft reset when local fork is present
And also when setting an undo limit, which works analogously. It was
caused by failing to increment the canvas state refcount when replaying
the local fork, which decrements it once. The function now indicates in
its name that it does so.

This also circumvents a bogus assertion botch when a local fork is
present while soft resetting or setting the undo depth.