Intuit background color when loading ORA and PSD
Many programs, such as Krita or GIMP, don't use a canvas background
color. Instead, they stick a single-colored layer at the bottom of the
stack, which they save into their ORA files. The PSD format doesn't
support background colors at all.

That's kind of annoying when loading them into Drawpile, since you end
up with a transparent background and this extraneous layer at the
bottom.

So now Drawpile will guess if the bottom layer is supposed to be a
background by checking if it's a regular layer with only one color, 100%
opacity and using the Normal blend mode. If so, that layer is turned
into the background color instead.

This only affects ORA files not saved by Drawpile itself. If it's one of
our own files, we don't do any of this guessery.