Rework the login flow to be less confusing
It now asks the user how they want to log in as the first step, rather
than taking a username and then guessing what they meant to do. This
fixes the common problem of someone wanting to join as a guest but
picking a registered username, then getting massively confused with
what's going on because they're being asked to log into an account they
don't have. Now they get to pick if they want to log in without an
account and if they choose a taken username, they're told to pick a
different one. The same goes for external and internal account clashes.

The user also now doesn't get booted when they enter an incorrect
password. This was already partially implemented, but both the server
and the client severed the connection when encounting a wrong password,
so it didn't actually work. Now it just lets you try again, rather than
having to restart the process from the beginning.