Allow taking current brush settings of other users
By opening the context menu on that user in the userlist next to chat
and selecting the "Take Current Brush" option. If that user has an
updated client, it will respond with the brush settings currently in use
and those will replace whichever brush settings are in the current slot.

This isn't a fully-featured brush sharing interface, but it's good
enough for many cases I think, where you just wanna try someone else's
brush real quick.