Allow using gesture instead of touch events
In the input preferences, users can now pick if they want to use touch
or gesture events. The gesture event handling hopefully works for stuff
like laptops and such.

Also slightly rejigs the input preferences layout to avoid having to
scroll down. The touch settings are now underneath the tablet ones and
the tablet tester has a label identifying it as such.