Shortcut login flow on servers without ext-auth
On servers that only allow guest logins and internal accounts, instead
of asking the user how they want to log in, they are immediately
forwarded to the guest login page. If they enter a name that's taken by
a server account, they'll be switched to the server account login page.

This is similar to the old login flow, except the error messages are
clearer. Servers with this kind of setup don't benefit from the new
login flow that asks the user how they want to log in, since they will
only have a very small amount of accounts, meaning that the chances of
accidentally entering a registered username are very slim and most users
will just want to log in as a guest anyway.