Fall back to static version when not in repo
If building outside of the repository, `git describe` will obviously
never manage to get a proper version, even if a git executable was
found. This caused the version to end up blank. It now emits a warning
and falls back to the statically determined PROJECT_VERSION in that case
instead.