Disallow links in host page session titles
Because some users accidentally paste invite links in there. When we
find a "://", we now turn off the host button and direct the user to go
to the join page instead. If someone *really* wants a URL in their
session title, they can still set it in the session settings after
hosting, this is only to discourage accidents.