Try to unlist session when stopping builtin server
There's still plenty of ways this can fail to work, but now a bit more
effort is made to unlist the session. If the server is closed, the
session will be unlisted before the server is outright deleted. When
closing the application while still connected, shutdown is delayed by a
second to let things settle gracefully.