Give timeline docker a proper minimum height
Making it show at least two tracks, which should be enough to also see
the "you haven't added any tracks yet" text. Otherwise the user may end
up with the timeline collapsed so far that they can only see the frame
indexes, which is confusing to a new user that may not know that
something is hidden from their view.