Handle setting of QT_VERSION in CMakeLists.txt
It used to work fine on its own, but on Arch Linux at least it seems
like the selection is now broken with the stock Qt cmake scripts. So now
we handle this instead by checking if a QT_VERSION is given on the
command line and looking for that Qt package explicitly.