Only apply user flags that are actually in effect
Rather than just saving and sending out the raw user flags received from
the authentication, some aspects of which may be denied by the server
configuration. In particular, this could lead to weird cases where a
user had the MOD flag from ext-auth, but is not actually a moderator
because the server denies ext-auth mods. Now any flags that aren't in
effect get filtered out.