Implement active drawing user counts for sessions
When announcing them and for the web admin API. A user is considered
actively drawing if a command message was received from them in the last
five minutes. The activity is not currently visible in the frontend and,
at the time of writing, the list server doesn't interpret this value.

Active drawing users are shown in session listings if the server
supports active user counts, otherwise they will be shown blank.
Filtering out inactive sessions is possible on the browse page.