Add server, session and user web permissions
To control who is allowed to join from a WebSocket connection, which
sessions allow WebSocket connections to them and which users are allowed
to toggle the WebSocket allowance of a session. By default, this is just
enabled for everyone, but it can be restricted to local user accounts by
giving them the WEB and WEBSESSION flags respectively, or it can be
pawned off to ext-auth. The session setting can be inspected and toggled
in the session settings dialog in the client or set through the admin
API.