Work around QTcpSocket's busted disconnect signals
The stateChanged and disconnected signals don't get emitted properly
when disconnecting the socket while it's establishing a connection. Now
we abort the socket in that case instead, which at least emits the
former, and treat a state change to the UnconnectedState as a
disconnect. This may cause multiple serverDisconnected signals to be
emitted, but WebSockets do that anyway, so we already handle that
behavior properly.