Fix flickering square when drawing on Windows
There's a visual bug on some systems when drawing with Windows Ink
enabled where a square region around the cursor flickers around when the
brush outline is disabled. This works around it by keeping the outline
graphics item always visible and retaining a bounding box, which I guess
causes the canvas to refresh under the cursor and prevents that flicker.