Don't double up disconnect notifications
Since it's possible to receive multiple of them, it would show both a
message at the top and a message box saying it couldn't connect to the
server, the latter of which is bogus. Now the message box is blocked by
the same expiry timer that prevents duplicate disconnect sound effects
being played.