Remove Rust from client again
It causes too many troubles with the WebAssembly build, since Rust
doesn't want to build that multithreaded. It can be worked around, but
it requires excessive hacking, building the standard library ourselves,
using unsupported build flags and a nightly Rust, none of which is
remotely guaranteed to be stable and not catch on fire if threads get
involved. So it's gone again, Rust is only used in the command-line
tools now. We can look at it again later, if they end up adding proper
multithreading support to it.