Add password-dependent web session setting
Meaning that joining from the web is automatically changed according to
if a password is set on the session or not, even if the user doesn't
have permission to adjust the setting otherwise. This lets server owners
allow joining only private sessions from the web, which reduces the
potential for abuse. Users with the WEBSESSION flag can override this,
as before.