Handle sessions running out of space better
There's now 1 MiB of emergency session space reserved for join, leave
and session owner messages. These are required to get a session reset
properly and are all just a few bytes in size (avatars are removed from
joins), so unless someone is intentionally causing a pathological
situation by joining and leaving millions of times, they will get
nowhere near filling up the emergency space. Server meta messages that
don't go into the emergency space get broadcast to all users directly.

The message that is sent when the session is out of space is now an
alert, rather than a regular message. It's also more informative,
pointing the user to where the reset option is, and it has a translation
key. Finally, the message no longer gets spammed for every single
message that couldn't be added to the history, instead it only gets sent
once every 10 seconds or when someone joins.